renaming store folder
